get_ext {rio}	R Documentation

Get File Type from Extension

Description

A utility function to retrieve the file type from a file extension (via its filename/path/URL)

Usage

get_ext(file)

Arguments

file
A character string containing a filename, file path, or URL.

Value

A characters string containing a file type recognized by rio.

[Package rio version 0.5.26 Index]