install_formats {rio}	R Documentation

Install rio's ‘Suggests’ Dependencies

Description

This function installs various ‘Suggests’ dependencies for rio that expand its support to the full range of support import and export formats. These packages are not installed or loaded by default in order to create a slimmer and faster package build, install, and load.

Usage

install_formats(...)

Arguments

...
Additional arguments passed to install.packages.

Value

NULL

[Package rio version 0.5.26 Index]