Post envelopes

The function upload a species distribution model's envelope with related information described as arguments to be stored in IO database.

post_envelopes(
  geom_ewkt,
  category,
  taxa_scientific_name,
  year,
  scenario = NULL,
  threshold_desc = NULL
)

Arguments

geom_ewkt
string. Polygon or Multipolygon geometry as Extended Well Known Text (EWKT) format with WGS84 as coordinate system (SRID = 4326)

category
string. One of either values current or projected. Describes if envelopes are modeled at current observations and environment variables or projected in future conditions.

taxa_scientific_name
string. Scientific name of organism for which the distribution is modeled.

year
integer. Reference year for current or projected distribution envelope.

scenario
string Optional. Name of the scenarion describing projected environment variables.

threshold_desc
string Optional. Name or short description for SDM thresholding methodology to generate a binary.

Value

NULL when successful

Examples

# Insert a small enveloppe for the newly discovered species `Binto binto`
response <- post_envelopes(
 geom_ewkt = "SRID=4326;POLYGON((-76.355 37.009,-76.350 37.010,-76.349 37.001,-76.354 37.001,-76.355 37.009))",
 category = "current",
 taxa_scientific_name = "Binto binto",
 year = 2022,
 scenario = NULL,
 threshold_desc = NULL
)