Convert from one file format to another

This function constructs a data frame from a data file using import and uses export to write the data to disk in the format indicated by the file extension.

convert(in_file, out_file, in_opts = list(), out_opts = list())

Arguments

in_file
A character string naming an input file.

out_file
A character string naming an output file.

in_opts
A named list of options to be passed to import.

out_opts
A named list of options to be passed to export.

Value

A character string containing the name of the output file (invisibly).

See also

Examples