Convert from one file format to another

This function constructs a data frame from a data file using import and uses export to write the data to disk in the format indicated by the file extension.

convert(in_file, out_file, in_opts = list(), out_opts = list())

Arguments

in_file
A character string naming an input file.

out_file
A character string naming an output file.

in_opts
A named list of options to be passed to import.

out_opts
A named list of options to be passed to export.

Value

A character string containing the name of the output file (invisibly).

See also

Examples

# create a file to convert
export(mtcars, "mtcars.dta")

# convert Stata to CSV and open converted file
convert("mtcars.dta", "mtcars.csv")

#> Warning: the condition has length > 1 and only the first element will be used

head(import("mtcars.csv"))

#> Warning: the condition has length > 1 and only the first element will be used

#>    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> 4 21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> 5 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> 6 18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

# correct an erroneous file format
export(mtcars, "mtcars.csv", format = "tsv")
convert("mtcars.csv", "mtcars.csv", in_opts = list(format = "tsv"))

#> Warning: the condition has length > 1 and only the first element will be used

# convert serialized R data.frame to JSON
export(mtcars, "mtcars.rds")
convert("mtcars.rds", "mtcars.json")

#> Warning: the condition has length > 1 and only the first element will be used

#> Loading required namespace: jsonlite

# cleanup
unlink("mtcars.csv")
unlink("mtcars.dta")
unlink("mtcars.rds")
unlink("mtcars.json")
# NOT RUN {
# convert from the command line:
Rscript -e "rio::convert('mtcars.dta', 'mtcars.csv')"
# }